实现回复功能，完善异常处理

+++ src/main/java/com/firstweb/demo/controller/QuestionController.java
@@ -17,7 +17,7 @@ public class QuestionController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id")Integer id,
+    public String question(@PathVariable(name = "id")long id,
                            Model model){
         QuestionPOJO questionPOJO=questionService.getById(id);
         //累加阅读数
