+++ src/main/java/com/firstweb/demo/service/QuestionService.java
@@ -64,7 +64,7 @@ public class QuestionService {
         return pagePOJO;
     }
 
-    public PagePOJO list(Integer userId, Integer page, Integer size) {
+    public PagePOJO list(Long userId, Integer page, Integer size) {
         PagePOJO pagePOJO = new PagePOJO();
         Integer totalPage;
         QuestionExample questionExample = new QuestionExample();
@@ -101,7 +101,7 @@ public class QuestionService {
         return pagePOJO;
     }
 
-    public QuestionPOJO getById(Integer id) {
+    public QuestionPOJO getById(Long id) {
         Question question = questionMapper.selectByPrimaryKey(id);
         if (question == null) {
             throw new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -139,7 +139,7 @@ public class QuestionService {
         }
     }
 
-    public void intView(Integer id) {
+    public void intView(Long id) {
         Question question = new Question();
         question.setId(id);
         question.setViewCount(1);
